fixup! clock: Add driver for ADI ADSP-SC5xx PLL
const char * const *parent_names is the correct and most precise way to
declare a pointer to an array of constant strings. The const on the
right side of the first asterisk * specifies that the strings themselves
are constant, and the const on the left side specifies that the pointers
in the array are also constant. This is the standard in Linux kernel
development for arrays of strings that won't be modified after
initialization.

void ** is the wrong type for an I/O memory offset. It's a pointer to a
pointer to void, which is used for generic double indirection.
Memory-mapped I/O (MMIO) offsets are typically declared as an unsigned
long or a simple u32, depending on the address bus width. unsigned long
is a standard way to represent a memory address or an offset on most
architectures.